+++ ChallengeCapi/DragonMain/Adapters/ListAdapter/ListCharacterGrillAdapter.swift
@@ -1,0 +1,19 @@
+//
+//  ListCharacterGrillAdapter.swift
+//  ChallengeCapi
+//
+//  Created by Juan Felipe Acevedo Ramirez on 16/04/24.
+//
+
+import UIKit
+
+class ListCharacterGrillAdapter: NSObject, ListCharacterAdapterProtocol {
+    
+    var datasource: [Any] = []
+    
+    func setCollectionView(_ collectionView: UICollectionView) {
+        <#code#>
+    }
+    
+    
+}