+++ ChallengeCapi/DragonMain/Adapters/ListAdapter/ListCharacterSimpleAdapter.swift
@@ -1,0 +1,18 @@
+//
+//  ListCharacterSimpleAdapter.swift
+//  ChallengeCapi
+//
+//  Created by Juan Felipe Acevedo Ramirez on 16/04/24.
+//
+
+import UIKit
+
+class ListCharacterSimpleAdapter: NSObject, ListCharacterAdapterProtocol {
+    var datasource: [Any] = []
+    
+    func setCollectionView(_ collectionView: UICollectionView) {
+        <#code#>
+    }
+    
+    
+}