+++ ChallengeCapi/DragonMain/Characters/CharacterCollectionViewCell.swift
@@ -1,0 +1,8 @@
+//
+//  CharacterCollectionViewCell.swift
+//  ChallengeCapi
+//
+//  Created by Juan Felipe Acevedo Ramirez on 17/04/24.
+//
+
+import Foundation