+++ ChallengeCapi/DragonMain/CharactesFavorites/FavoriteCaracterCollectionViewCell.swift
@@ -1,0 +1,8 @@
+//
+//  FavoriteCaracterCollectionViewCell.swift
+//  ChallengeCapi
+//
+//  Created by Juan Felipe Acevedo Ramirez on 17/04/24.
+//
+
+import Foundation