+++ ChallengeCapi/TabBarController/TapBarControllerViewController.swift
@@ -1,0 +1,15 @@
+//
+//  TapBarControllerViewController.swift
+//  ChallengeCapi
+//
+//  Created by Juan Felipe Acevedo Ramirez on 16/04/24.
+//
+
+import UIKit
+
+class TapBarControllerViewController: UITabBarController {
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+    }
+}